chore: relocate products directives to submodule
This change involves moving the directive definitions to their
respective "submodules". This is an initial step toward modularizing the
module, ensuring a clear separation of concerns between the different
products.
